+++ src/services/ai-server/ai-server/ai.server.resolver.mutations.ts
@@ -1,6 +1,6 @@
-import { UseGuards } from '@nestjs/common';
-import { Resolver, Mutation, Args } from '@nestjs/graphql';
-import { CurrentUser } from '@src/common/decorators';
+import { Inject, LoggerService, UseGuards } from '@nestjs/common';
+import { Resolver, Mutation, Args, Field, ObjectType } from '@nestjs/graphql';
+import { CurrentUser, Profiling } from '@src/common/decorators';
 import { GraphqlGuard } from '@core/authorization';
 import { AgentInfo } from '@core/authentication.agent.info/agent.info';
 import { AuthorizationService } from '@core/authorization/authorization.service';
@@ -12,6 +12,21 @@ import { AiPersonaServiceService } from '../ai-persona-service/ai.persona.servic
 import { AiPersonaServiceAuthorizationService } from '../ai-persona-service/ai.persona.service.authorization';
 import { CreateAiPersonaServiceInput } from '../ai-persona-service/dto/ai.persona.service.dto.create';
 import { IAiPersonaService } from '../ai-persona-service/ai.persona.service.interface';
+import { ConfigurationTypes } from '@common/enums';
+import { Space } from '@domain/space/space/space.entity';
+import { SpaceIngestionPurpose } from '@services/infrastructure/event-bus/commands';
+import { ChromaClient } from 'chromadb';
+import { ConfigService } from '@nestjs/config';
+import { InjectEntityManager } from '@nestjs/typeorm';
+import { PlatformAuthorizationPolicyService } from '@platform/authorization/platform.authorization.policy.service';
+import { EntityManager } from 'typeorm';
+import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
+
+@ObjectType('MigrateEmbeddings')
+class IMigrateEmbeddingsResponse {
+  @Field(() => Boolean, { description: 'Result from the mutation execution.' })
+  success!: boolean;
+}
 
 @Resolver()
 export class AiServerResolverMutations {
@@ -20,8 +35,141 @@ export class AiServerResolverMutations {
     private aiServerService: AiServerService,
     private aiServerAuthorizationService: AiServerAuthorizationService,
     private aiPersonaServiceService: AiPersonaServiceService,
-    private aiPersonaServiceAuthorizationService: AiPersonaServiceAuthorizationService
+    private aiPersonaServiceAuthorizationService: AiPersonaServiceAuthorizationService,
+    private platformAuthorizationService: PlatformAuthorizationPolicyService,
+    @InjectEntityManager('default')
+    private entityManager: EntityManager,
+    private config: ConfigService,
+    @Inject(WINSTON_MODULE_NEST_PROVIDER)
+    private logger: LoggerService
   ) {}
+
+  @UseGuards(GraphqlGuard)
+  @Mutation(() => IMigrateEmbeddingsResponse, {
+    description: 'Deletes collections nameID-...',
+  })
+  @Profiling.api
+  async cleanupCollections(
+    @CurrentUser() agentInfo: AgentInfo
+  ): Promise<IMigrateEmbeddingsResponse> {
+    const platformAuthorization =
+      await this.platformAuthorizationService.getPlatformAuthorizationPolicy();
+    this.authorizationService.grantAccessOrFail(
+      agentInfo,
+      platformAuthorization,
+      AuthorizationPrivilege.PLATFORM_ADMIN,
+      'User not authenticated to migrate embeddings'
+    );
+
+    const vectorDb = this.config.get(ConfigurationTypes.PLATFORM).vector_db;
+
+    const chroma = new ChromaClient({
+      path: `http://${vectorDb.host}:${vectorDb.port}`,
+    });
+
+    // get all collections
+    const collections = await chroma.listCollections();
+    for (const collection of collections) {
+      // extract collection identifier and purpose
+      const [, nameID, purpose] =
+        collection.name.match(/(.*)-(knowledge|context)/) || [];
+
+      // if the identifier is of UUID format, skip it
+      if (
+        /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/.test(
+          nameID
+        )
+      ) {
+        continue;
+      }
+
+      // get the space by nameID
+      const space = await this.entityManager.findOne(Space, {
+        where: { nameID: nameID },
+      });
+
+      // if the space doesn't exit skip the colletion
+      if (!space) {
+        this.logger.warn(
+          `Space with nameID ${nameID} does't exist but ${collection.name} is still in Chroma`
+        );
+        continue;
+      }
+
+      // check if there is UUID identified collection for this space
+      const uuidCollection = await chroma.getCollection({
+        name: `${space.id}-${purpose}`,
+      });
+
+      // if so, delete the nameID one
+      if (uuidCollection) {
+        await chroma.deleteCollection(collection);
+      }
+    }
+
+    return { success: true };
+  }
+  @UseGuards(GraphqlGuard)
+  @Mutation(() => IMigrateEmbeddingsResponse, {
+    description: 'Copies collections nameID-... into UUID-...',
+  })
+  @Profiling.api
+  async copyCollections(
+    @CurrentUser() agentInfo: AgentInfo
+  ): Promise<IMigrateEmbeddingsResponse> {
+    const platformAuthorization =
+      await this.platformAuthorizationService.getPlatformAuthorizationPolicy();
+    this.authorizationService.grantAccessOrFail(
+      agentInfo,
+      platformAuthorization,
+      AuthorizationPrivilege.PLATFORM_ADMIN,
+      'User not authenticated to migrate embeddings'
+    );
+
+    const vectorDb = this.config.get(ConfigurationTypes.PLATFORM).vector_db;
+
+    const chroma = new ChromaClient({
+      path: `http://${vectorDb.host}:${vectorDb.port}`,
+    });
+    // get all chroma collections
+    const collections = await chroma.listCollections();
+
+    for (const collection of collections) {
+      // extract collection identifier and purpose
+      const [, nameID, purpose] =
+        collection.name.match(/(.*)-(knowledge|context)/) || [];
+
+      // if the identifier is of UUID format, skip it
+      if (
+        /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}/.test(
+          nameID
+        )
+      ) {
+        continue;
+      }
+      // get the space by nameID
+      const space = await this.entityManager.findOne(Space, {
+        where: { nameID: nameID },
+      });
+
+      // if the space doesn't exit skip the colletion
+      if (!space) {
+        this.logger.warn(
+          `Space with nameID ${nameID} does't exist but ${collection.name} is still in Chroma`
+        );
+        continue;
+      }
+
+      // ask the AI server to ingest the space again;
+      // the ingest space service uses the UUID as collection identifier now
+      this.aiServerService.ensureSpaceIsUsable(
+        space.id,
+        purpose as SpaceIngestionPurpose
+      );
+    }
+
+    return { success: true };
+  }
 
   @UseGuards(GraphqlGuard)
   @Mutation(() => IAiServer, {
